decouple funcitons
