+++ core/src/debt.ts
@@ -1,35 +1,41 @@
-import type { DebtAuction, DebtAuctionActive, DebtAuctionTransaction, Notifier } from './types';
-import { getCompensationAuctionTransactionFees } from './compensationAuction';
+import type { DebtAuctionActive, DebtAuctionTransaction } from './types';
 import BigNumber from './bignumber';
 import getContract from './contracts';
 import { RAD_NUMBER_OF_DIGITS, WAD_NUMBER_OF_DIGITS } from './constants/UNITS';
-import executeTransaction from './execute';
+import { Contract } from 'ethers';
+import { MKR_NUMBER_OF_DIGITS, RAD, WAD } from './constants/UNITS';
+import memoizee from 'memoizee';
+import getNetworkDate from './date';
 import {
-    collectCompensationAuction,
-    getActiveCompensationAuctionOrUndefined,
-    getCompensationAuctionBidIncreaseCoefficient,
-    getCompensationAuctionLastIndex,
-    getMarketPriceDaiToMkr,
-    restartCompensationAuction,
-} from './compensationAuction';
+    CompensationAuctionBase,
+    DebtAuction,
+    Notifier,
+    SurplusAuction,
+    CompensationAuctionTransactionFees,
+} from './types';
+import { getEarliestDate } from './helpers/getEarliestDate';
+import executeTransaction from './execute';
+import { convertMkrToDai } from './calleeFunctions/helpers/uniswapV3';
+import { getGasPriceForUI } from './gas';
+import { getMarketPrice } from './calleeFunctions';
 
 const CONTRACT = 'MCD_FLOP';
 export const getNextMaximumMkrReceived = async (
     network: string,
     debtAuction: DebtAuctionActive
 ): Promise<BigNumber> => {
-    const increaseCoefficient = await getCompensationAuctionBidIncreaseCoefficient(network, CONTRACT);
+    const increaseCoefficient = await getDebtAuctionBidIncreaseCoefficient(network, CONTRACT);
     return debtAuction.receiveAmountMKR.dividedBy(increaseCoefficient);
 };
 
 export const fetchActiveDebtAuctions = async function (network: string): Promise<DebtAuctionActive[]> {
     const contract = await getContract(network, CONTRACT);
-    const auctionLastIndex = await getCompensationAuctionLastIndex(contract);
+    const auctionLastIndex = await getDebtAuctionLastIndex(contract);
 
     const surplusAuctions: DebtAuctionActive[] = [];
     let currentSurplusAuction;
     for (let i = auctionLastIndex; i > 0; i--) {
-        const currentAuction = await getActiveCompensationAuctionOrUndefined(network, i, CONTRACT);
+        const currentAuction = await getActiveDebtAuctionOrUndefined(network, i);
         if (!currentSurplusAuction) {
             break;
         }
@@ -43,7 +49,7 @@ export const restartDebtAuction = async function (
     auctionIndex: number,
     notifier?: Notifier
 ): Promise<void> {
-    await restartCompensationAuction(network, auctionIndex, CONTRACT, notifier);
+    await executeTransaction(network, 'MCD_FLOP', 'tick', [auctionIndex], { notifier });
 };
 
 export const bidToDebtAuction = async function (
@@ -52,9 +58,7 @@ export const bidToDebtAuction = async function (
     acceptableReceivedAmount: BigNumber,
     notifier?: Notifier
 ) {
-    const auction = (await getActiveCompensationAuctionOrUndefined(network, auctionIndex, CONTRACT)) as
-        | DebtAuctionActive
-        | undefined;
+    const auction = (await getActiveDebtAuctionOrUndefined(network, auctionIndex)) as DebtAuctionActive | undefined;
     if (!auction || !auction.receiveAmountMKR) {
         throw new Error('Did not find the auction to bid on.');
     }
@@ -66,7 +70,11 @@ export const bidToDebtAuction = async function (
     await executeTransaction(network, CONTRACT, 'dent', transactionParameters, { notifier });
 };
 export const collectDebtAuction = async (network: string, auctionIndex: number, notifier?: Notifier) => {
-    await collectCompensationAuction(network, auctionIndex, CONTRACT, notifier);
+    const auction = await getActiveDebtAuctionOrUndefined(network, auctionIndex);
+    if (!auction || auction.state !== 'ready-for-collection') {
+        throw new Error('Did not find the auction to collect.');
+    }
+    await executeTransaction(network, 'MCD_FLOP', 'deal', [auctionIndex], { notifier });
 };
 export const enrichDebtAuction = async (
     network: string,
@@ -76,7 +84,7 @@ export const enrichDebtAuction = async (
     const unitPrice = auction.bidAmountDai.div(auction.receiveAmountMKR);
     const marketUnitPrice = await getMarketPriceDaiToMkr(network, auction.receiveAmountMKR);
     const marketUnitPriceToUnitPriceRatio = unitPrice.minus(marketUnitPrice).dividedBy(marketUnitPrice);
-    const fees = await getCompensationAuctionTransactionFees(network);
+    const fees = await getDebtAuctionTransactionFees(network);
     return {
         ...auction,
         ...fees,
@@ -95,4 +103,145 @@ export const enrichDebtAuctions = async (network: string, auctions: DebtAuction[
         return enrichDebtAuction(network, auction);
     });
     return await Promise.all(auctionsWithNextMinimumBidsPromises as Promise<DebtAuctionTransaction>[]);
+};
+
+export const getDebtAuctionLastIndex = async (contract: Contract): Promise<number> => {
+    const auctionsQuantityBinary = await contract.kicks();
+    return new BigNumber(auctionsQuantityBinary._hex).toNumber();
+};
+
+const _getDebtAuctionBidIncreaseCoefficient = async (
+    network: string,
+    contractName: 'MCD_FLAP' | 'MCD_FLOP'
+): Promise<BigNumber> => {
+    const contract = await getContract(network, contractName);
+    const auctionsQuantityBinary = await contract.beg();
+    return new BigNumber(auctionsQuantityBinary._hex).shiftedBy(-MKR_NUMBER_OF_DIGITS);
+};
+
+export const getDebtAuctionBidIncreaseCoefficient = memoizee(_getDebtAuctionBidIncreaseCoefficient, {
+    promise: true,
+    length: 3,
+});
+
+const getDebtAuctionState = async (network: string, earliestEndDate: Date, greatestBid: number) => {
+    const isBidExpired = (await getNetworkDate(network)) > earliestEndDate;
+    const haveBids = greatestBid !== 0;
+    if (haveBids) {
+        if (isBidExpired) {
+            return 'ready-for-collection';
+        }
+        return 'have-bids';
+    }
+    if (isBidExpired) {
+        return 'requires-restart';
+    }
+    return 'just-started';
+};
+
+const _getContractAuctionDuration = async (network: string, contractName: 'MCD_FLAP' | 'MCD_FLOP') => {
+    const contract = await getContract(network, contractName);
+    return await contract.tau();
+};
+const getContractAuctionDuration = memoizee(_getContractAuctionDuration, {
+    maxAge: 24 * 60 * 60 * 1000,
+    promise: true,
+    length: 1,
+});
+
+export const fetchDebtAuctionByIndex = async (
+    network: string,
+    auctionIndex: number,
+    contractName: 'MCD_FLAP' | 'MCD_FLOP'
+): Promise<SurplusAuction | DebtAuction> => {
+    const contract = await getContract(network, contractName);
+    const auctionData = await contract.bids(auctionIndex);
+    const isAuctionCollected = new BigNumber(auctionData.end).eq(0);
+    const auctionDuration = await getContractAuctionDuration(network, contractName);
+    const fetchedAt = new Date();
+    const baseAuctionInfo: CompensationAuctionBase = {
+        network,
+        id: auctionIndex,
+        fetchedAt,
+    };
+
+    if (isAuctionCollected) {
+        const auctionLastIndex = await getDebtAuctionLastIndex(contract);
+        if (auctionLastIndex < auctionIndex) {
+            throw new Error('No active auction exists with this id');
+        }
+        return { ...baseAuctionInfo, state: 'collected' };
+    }
+
+    const auctionEndDate = new Date(auctionData.end * 1000);
+    const auctionStartDate = new Date(auctionEndDate.getTime() - auctionDuration * 1000);
+    const bidEndDate = auctionData.tic ? new Date(auctionData.tic * 1000) : undefined;
+    const earliestEndDate = bidEndDate ? getEarliestDate(auctionEndDate, bidEndDate) : auctionEndDate;
+    const state = await getDebtAuctionState(network, earliestEndDate, auctionData.tic);
+
+    return {
+        ...baseAuctionInfo,
+        earliestEndDate,
+        bidAmountMKR: new BigNumber(auctionData.bid._hex).div(WAD),
+        receiveAmountDAI: new BigNumber(auctionData.lot._hex).div(RAD),
+        receiverAddress: auctionData.guy,
+        auctionEndDate,
+        bidEndDate,
+        state,
+        fetchedAt,
+        auctionStartDate,
+    };
+};
+
+export const getActiveDebtAuctionOrUndefined = async (network: string, auctionIndex: number) => {
+    const auction = await fetchDebtAuctionByIndex(network, auctionIndex, 'MCD_FLOP');
+    if (auction.state === 'collected') {
+        return;
+    }
+    return auction;
+};
+
+export const getMarketPriceMkr = async function (network: string, mkrAmount: BigNumber): Promise<BigNumber> {
+    try {
+        return new BigNumber(1).div((await convertMkrToDai(network, mkrAmount)).div(mkrAmount));
+    } catch (error) {
+        return new BigNumber(NaN);
+    }
+};
+
+export const getMarketPriceDaiToMkr = async function (network: string, mkrAmount: BigNumber): Promise<BigNumber> {
+    try {
+        return (await convertMkrToDai(network, mkrAmount)).div(mkrAmount);
+    } catch (error) {
+        return new BigNumber(NaN);
+    }
+};
+
+export const getDebtAuctionTransactionFees = async function (
+    network: string
+): Promise<CompensationAuctionTransactionFees> {
+    const gasPrice = await getGasPriceForUI(network);
+    const exchangeRate = await getMarketPrice(network, 'ETH');
+
+    const restartTransactionFeeEth = gasPrice.multipliedBy(80563);
+    const allowanceTransactionFeeEth = gasPrice.multipliedBy(48373);
+    const bidTransactionFeeEth = gasPrice.multipliedBy(85181);
+    const collectTransactionFeeEth = gasPrice.multipliedBy(94114);
+    const authTransactionFeeEth = gasPrice.multipliedBy(48356);
+    const combinedBidFeesEth = bidTransactionFeeEth.plus(collectTransactionFeeEth);
+
+    return {
+        restartTransactionFeeEth,
+        restartTransactionFeeDai: restartTransactionFeeEth.multipliedBy(exchangeRate),
+        allowanceTransactionFeeEth,
+        allowanceTransactionFeeDai: allowanceTransactionFeeEth.multipliedBy(exchangeRate),
+        bidTransactionFeeEth,
+        bidTransactionFeeDai: bidTransactionFeeEth.multipliedBy(exchangeRate),
+        collectTransactionFeeEth,
+        collectTransactionFeeDai: collectTransactionFeeEth.multipliedBy(exchangeRate),
+        authTransactionFeeEth,
+        authTransactionFeeDai: authTransactionFeeEth.multipliedBy(exchangeRate),
+        combinedBidFeesEth,
+        combinedBidFeesDai: combinedBidFeesEth.multipliedBy(exchangeRate),
+    };
 };
